+++ src/Game/SimpleLevelObject.h
@@ -35,7 +35,7 @@ public:
 		return (SimpleLevelObjectClass*)class_;
 	}
 
-	virtual bool getTexture(size_t textureNumber, TextureInfo& ti) const;
+	virtual bool getTexture(uint32_t textureNumber, TextureInfo& ti) const;
 
 	virtual bool getNumberProp(const std::string_view prop, Number32& value) const noexcept
 	{
